Slot render: use supplied pose stack instead of creating a new one
Creating a new pose stack breaks any mod that move the hotbar, such as
Auto HUD or Raised. The item border would always be drawn at the vanilla
hotbar position instead of where the item gets actually rendered. By
using the pose stack as it was handed to the method, other mods' matrix
translations will be respected. This properly anchors the border to the
rendered item.